+++ src/encoder_formats/ffmpeg_format.ml
@@ -21,14 +21,19 @@
  *****************************************************************************)
 
 type opt_val = [ `String of string | `Int of int | `Float of float ]
+type output = [ `Stream | `Url of string ]
 type opts = (string, opt_val) Hashtbl.t
 
 type t = {
-  format : string;
-  codec : string;
+  format : string option;
+  output : output;
   channels : int;
   samplerate : int Lazy.t;
-  options : opts;
+  audio_codec : string option;
+  audio_opts : opts;
+  video_codec : string option;
+  video_opts : opts;
+  other_opts : opts;
 }
 
 let string_of_options options =
@@ -45,7 +50,36 @@ let string_of_options options =
        options [])
 
 let to_string m =
-  let opts = string_of_options m.options in
-  Printf.sprintf "%%fmpeg(format=%S,codec=%S,ac=%d,ar=%d%s)" m.format m.codec
-    m.channels (Lazy.force m.samplerate)
+  let audio_codec =
+    match m.audio_codec with
+      | None -> ""
+      | Some c -> Printf.sprintf ",audio_codec=%S" c
+  in
+  let video_codec =
+    match m.video_codec with
+      | None -> ""
+      | Some c -> Printf.sprintf ",video_codec=%S" c
+  in
+  let opts = string_of_options m.other_opts in
+  let audio_opts =
+    if Hashtbl.length m.audio_opts > 0 then
+      Printf.sprintf ",%%audio(%s)" (string_of_options m.audio_opts)
+    else ""
+  in
+  let video_opts =
+    if Hashtbl.length m.video_opts > 0 then
+      Printf.sprintf ",%%video(%s)" (string_of_options m.video_opts)
+    else ""
+  in
+  let format =
+    match m.format with Some f -> Printf.sprintf ",format=%S" f | None -> ""
+  in
+  let output =
+    match m.output with
+      | `Stream -> ""
+      | `Url path -> Printf.sprintf ",url=%S" path
+  in
+  Printf.sprintf "%%fmpeg(channels=%d,samplerate=%d%s%s%s%s%s%s%s)" m.channels
+    (Lazy.force m.samplerate) audio_codec audio_opts video_codec video_opts
     (if opts = "" then "" else Printf.sprintf ",%s" opts)
+    format output